+++ GPUMultiFIlter/Shaders/InkXTo.glsl
@@ -2,7 +2,7 @@ varying highp vec2 textureCoordinate;
 uniform sampler2D FixInkMap;
 uniform sampler2D SinkInkMap;
 uniform sampler2D velDen;
-uniform lowp float bEvaporToDisapper; // false = 0, true = 1
+uniform highp float bEvaporToDisapper; // false = 0, true = 1
 
 void main(void)
 {
@@ -10,12 +10,12 @@ void main(void)
  
  highp vec4 sink = texture2D(SinkInkMap, Tex0);
  highp vec4 ix0 = texture2D(FixInkMap, Tex0);
- lowp vec4 ix_new = sink + ix0;
+ highp vec4 ix_new = sink + ix0;
  ix_new = vec4(ix_new.xyz, (ix_new.x + ix_new.y + ix_new.z) / 3.0);
  
  if (bEvaporToDisapper == 1.0)
  {
-     mediump vec4 a = texture2D(velDen, Tex0);
+     highp vec4 a = texture2D(velDen, Tex0);
      if (a.z <= 0.001)
          gl_FragColor = vec4(ix_new.rgb * (a.z * 990.0), ix_new.a);
      else
